fix: refactor business unit and checking credentials page
