+++ src/design/feedback/noResultsMessage/styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { tokens } from "@design/tokens";
+
+const StyledNoResults = styled.div`
+  margin: ${tokens.spacing.s200} ${tokens.spacing.s0};
+`;
+
+export { StyledNoResults };
