+++ src/hooks/selectBusinessUnits/useBusinessUnits/index.ts
@@ -1,7 +1,9 @@
-import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortal/IBusinessUnitsPortalStaff";
-import { AuthAndPortalData } from "@context/authAndPortalDataProvider";
 import { useContext, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { useMediaQueries } from "@inubekit/inubekit";
+
+import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortal/IBusinessUnitsPortalStaff";
+import { AuthAndPortalData } from "@context/authAndPortalDataProvider";
 import { IBusinessUnitstate } from "@ptypes/selectBusinessUnits/outlets/businessUnit/IBusinessUnitstate";
 
 const useBusinessUnits = (businessUnits: IBusinessUnitsPortalStaff[]) => {
@@ -58,9 +60,19 @@ const useBusinessUnits = (businessUnits: IBusinessUnitsPortalStaff[]) => {
     });
   }
 
+  const {
+    "(max-width: 532px)": screenMobile,
+    "(max-height: 1000px)": screenTablet,
+  }: Record<string, boolean> = useMediaQueries([
+    "(max-width: 532px)",
+    "(max-height: 1000px)",
+  ]);
+
   return {
     search,
     businessUnitLocal,
+    screenMobile,
+    screenTablet,
     handleSearchChange,
     handleChange,
     handleSubmit,
