+++ src/pages/login/checkingCredentials/index.tsx
@@ -1,9 +1,9 @@
-import { useNavigate } from "react-router-dom";
-import { useCallback, useContext, useEffect } from "react";
+import { useContext } from "react";
 
 import { IBusinessUnitsPortalStaff } from "@ptypes/staffPortal/IBusinessUnitsPortalStaff";
-import { CheckingCredentialsUI } from "./interface";
 import { AuthAndPortalData } from "@context/authAndPortalDataProvider";
+import { useCheckingCredentials } from "@hooks/authentication/useCheckingCredentials";
+import { CheckingCredentialsUI } from "./interface";
 
 interface ICheckingCredentials {
   businessUnits: IBusinessUnitsPortalStaff[];
@@ -11,42 +11,10 @@ interface ICheckingCredentials {
 
 function CheckingCredentials(props: ICheckingCredentials) {
   const { businessUnits } = props;
-  const navigate = useNavigate();
+
   const { appData, setBusinessUnitSigla } = useContext(AuthAndPortalData);
 
-  const selectedBusinessUnit = () => {
-    const selected = businessUnits[0];
-    setBusinessUnitSigla(JSON.stringify(selected));
-    navigate("/selectBusinessUnit/loading-app");
-  };
-
-  const checkCredentials = useCallback(() => {
-    try {
-      if (!appData) {
-        return;
-      }
-
-      if (appData) {
-        if (!businessUnits || businessUnits.length === 0) {
-          navigate("/selectBusinessUnit/error/not-related-businessUnits");
-        } else if (businessUnits.length === 1) {
-          selectedBusinessUnit();
-        } else {
-          navigate(`/selectBusinessUnit/businessUnits`);
-        }
-      } else {
-        navigate("/selectBusinessUnit/error/not-available");
-      }
-    } catch (error) {
-      console.info(error);
-      navigate("/selectBusinessUnit/error/not-available");
-    }
-  }, [appData, navigate, businessUnits]);
-
-  useEffect(() => {
-    const timer = setTimeout(() => checkCredentials(), 2000);
-    return () => clearTimeout(timer);
-  }, [checkCredentials]);
+  useCheckingCredentials(businessUnits, appData, setBusinessUnitSigla);
 
   return <CheckingCredentialsUI />;
 }
